odonto-gpt-agno-service/requirements.txt
# Core Dependencies
agno>=2.0.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.23

# OpenAI & AI
openai>=1.6.0
anthropic>=0.7.0

# Research & Scientific Papers
arxiv>=2.1.0
pymed>=0.9.0

# HTTP & API
httpx>=0.25.0
python-multipart>=0.0.6

# CORS & Middleware
python-jose[cryptography]>=3.3.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
